Moved all test helper functions to a separate file.
